+++ Runtime/Scripts/Base/Raycasting.cs
@@ -20,8 +20,7 @@ namespace Kokowolo.Utilities
         /************************************************************/
         #region Functions
 
-        public static bool RaycastToDestinationPoint(Vector3 origin, Vector3 destination, out RaycastHit hitInfo, 
-            LayerMask layerMask)
+        public static bool RaycastToDestinationPoint(Vector3 origin, Vector3 destination, out RaycastHit hitInfo, LayerMask layerMask)
         {
             Vector3 direction = (destination - origin).normalized;
             float maxDistance = Vector3.Distance(origin, destination);
@@ -32,7 +31,7 @@ namespace Kokowolo.Utilities
             float maxDistance = Mathf.Infinity, Camera camera = null)
         {
             if (!camera) camera = Camera.main;
-            Ray ray = camera.ScreenPointToRay(BaseInputManager.GetMouseScreenPoint());
+            Ray ray = camera.ScreenPointToRay(InputManager.GetMouseScreenPoint());
             if (Physics.Raycast(ray, out hitInfo, maxDistance, layerMask))
             {
                 return true;